+++ packages/fighting-design/dialog/src/dialog.vue
@@ -8,7 +8,6 @@
 
   const prop = defineProps(Props)
   const visible = defineModel<boolean>('visible', {
-    required: true,
     default: false,
     type: Boolean
   })