+++ packages/fighting-design/input/src/input.vue
@@ -13,7 +13,6 @@
 
   const prop = defineProps(Props)
   const modelValue = defineModel<string | number>({
-    required: true,
     default: '',
     type: [String, Number]
   })