+++ packages/fighting-design/tree/components/index.ts
@@ -1,0 +1,1 @@
+export { default as FTreeItem } from './tree-item/index.vue'