+++ docs/.vitepress/config/sidebar.ts
@@ -106,7 +106,8 @@ export const sidebar = {
         { text: 'Dialog 对话框', link: '/components/dialog' },
         { text: 'Drawer 抽屉', link: '/components/drawer' },
         { text: 'Tooltip 消息提示', link: '/components/tooltip' },
-        { text: 'Swap 切换', link: '/components/swap' }
+        { text: 'Swap 切换', link: '/components/swap' },
+        { text: 'Confirm Box 确认框', link: '/components/confirm-box' }
       ]
     },
     {