+++ start/src/App.vue
@@ -1,21 +1,3 @@
-<script lang="ts" setup>
-  import { FConfirmBox } from 'fighting-design'
+<script lang="ts" setup></script>
 
-  const open = (): void => {
-    FConfirmBox({
-      title: '标题',
-      content: '这是内容',
-      onConfirm: () => {
-        return new Promise(res => {
-          setTimeout(() => {
-            res('123')
-          }, 2000)
-        })
-      }
-    })
-  }
-</script>
-
-<template>
-  <f-button :on-click="open">提示</f-button>
-</template>
+<template></template>