+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - 废除 `f-infinite-scrolling` 组件 `loading` 配置项
 - 重构 `f-infinite-scrolling` 组件
 - 重构 `f-tabs` 组件
+- 废除 `f-infinite-scrolling` 组件 `on-scroll-end` 参数类型修改
 
 ## 0.52.0 (2023-07-14)