+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+- 优化 `f-drawer` `f-dialog` 组件样式
+
 ## 0.60.0 (2023-08-04)
 
 - 优化 `f-radio` 组件样式